Removed PHP and Perl APIs
The included PHP and Perl APIs did not get actually maintained and
especially never used by anyone. They still were on the API level of
CUPS 1.2. And the Perl part was not even supported by the build system
(only the scripting/perl/ directory in EXTRA_DIST to get it into the
source tarball).

Therefore we do not take it over into cups-filters 2.x and drop it
here.