resolve_callback()/resolver_wrapper(): New thread only when printer found
We move the check which resolver event we have (found/failure) already
in the main thread (resolver_wrapper()) and launch a new thread only
if we have found a new printer and have to investigate whether to add
a queue for it or not. resolve_callback() only initiates this
investigation now.

This way we do not need to pass the resolver data structure (of type
AvahiServiceResolver*) into the new thread, which caused segfaults.

Also corrected some memory freeing when a printer data structure is
deleted (in create_remote_printer_entry()), but this has not caused a
segfault in the recent tests.