Prefer sending jobs in Apple Raster instead of in PDF
If a destination printer supports both PDF and Apple Raster, and if it
is not a remote CUPS queue, prefer sending the job in Apple Raster, as
printers print this more reliably. PDF interpreters in printers are
often slow or can easily have bugs.

Should fix https://bugs.launchpad.net/bugs/2014976