run-tests.sh: Let emulated printers support PDF input
To test that cups-browsed prefers Apple Raster when the printer
supports both PDF and Apple Raster as input format, we let the
printers emulated by ippeveprinter also support PDF. We are already
checking whether the output is actually Apple Raster, so if it would
be PDF, the test script fails.

Also updated the log messages appropriately.

This reproduces https://bugs.launchpad.net/bugs/2014976 where the
destination printer gets slow due to receiving PDF and not Apple
Raster with cups-browsed 2.0b4. The bug got fixed in

    https://github.com/OpenPrinting/cups-browsed/commit/3eb66da94d

Earlier versions of cups-browsed now will fail the test as with them
the emulated printer receives PDF and not Apple Raster.