Do not record *-default IPP attributes of local CUPS queues
Many of the *-default IPP attributes represent properties
already covered by the PPD option defaults which we also record.

I addition, there is also the problematic print-quality-default
attribute. IPP reports print-quality as "draft", "normal", and "high"
while CUPS only accepts "3", "4", and "5" (and NOT "draft", "normal",
and "high") and if an invalid value is supplied, print-quality-default
is set to 0 and as 0 is also invalid, jobs do not get printed. Now,
not touching print-quality-default this does not happen for
cups-browsed's local queues anymore and they reliably print.