Use description/location from server if available, otherwise from client
Fixes #323

When we create a local queue we first check whether we actually got
description and location strings from the remote server/printer, if
they are empty we do not set empty strings but use the IPP attributes
saved locally for our local queue.

This way, if the server does not provide description/location and the
user sets their own, that one is conserved through reboots and daemon
restarts.

Thanks, Edgar Fuß, for the hint to make sure both server-provided and
local-user-defined description/location work.