Do not generate PPD for remote raw queues
Currently, we generate PPD for remote raw queues if they are acquired by
polling the CUPS server via BrowsePoll. PPD generation will get filters
applied, which remote server does not take into account and printer
ready files become a different format.

It was discovered when user was passing ZPL file on the client expecting
it will arrive to the server as it is, but PPD generated by cups-browsed
cause ZPL file to be converted to PDF.