Fix uninitialized make_model in create_queue()
Fixes #42

At 2 points the string buffer for make_model got malloced but not
initialized by putting a terminating zero to its beginning.

At the same points sizeof() was applied to the pointer to the buffer
rsulting in a 7-byte limit and strncpy was used which does not put a
terminating zero when the string copied is too long for the given
limit (which was always the case). No we use an explicit number for
the limit and strncat which always zero-terminates.

Thanks, Solar Designer, for describing bug and solution so well in
your report.